HOME-82 - Implement relay toggler
