+++ client/models/agents/queries.js
@@ -1,6 +1,31 @@
 import _ from 'lodash';
 import * as selectors from './selectors';
 
+export const getTemperatures = (agent) => {
+  const { data } = agent;
+  return data.temperature;
+};
+
+export const getTemperature = (agent) => {
+  const temperatures = getTemperatures(agent);
+  return _.first(temperatures);
+};
+
+export const getMotion = (agent) => {
+  const { data } = agent;
+  return data.presence;
+};
+
+export const getGas = (agent) => {
+  const { data } = agent;
+  return data.gas;
+};
+
+export const getTimes = (agent) => {
+  const { data } = agent;
+  return data.time;
+};
+
 export const getTicks = (times, values) => _.map(values, (value, index) => ({
   time: new Date(times[index]),
   value,
@@ -11,17 +36,12 @@ export const getAgentById = (state, id) => {
   return _.find(agents, { id });
 };
 
-export const getTemperature = (agent) => {
-  const temperatures = selectors.getTemperatures(agent);
-  return _.first(temperatures);
-};
-
 export const isMotion = (agent) => {
-  const presence = selectors.getMotion(agent);
+  const presence = getMotion(agent);
   return _.reduce(presence, (acc, val) => acc || Number(val), 0);
 };
 
 export const isGas = (agent) => {
-  const gas = selectors.getGas(agent);
+  const gas = getGas(agent);
   return _.reduce(gas, (acc, val) => acc || Number(val), 0);
 };
