+++ client/models/agents/sagas.js
@@ -1,8 +1,7 @@
 import _ from 'lodash';
 import { delay } from 'redux-saga';
-import { put, call, fork, takeEvery } from 'redux-saga/effects';
+import { put, call } from 'redux-saga/effects';
 import * as actions from './actions';
-import * as actionTypes from './actionTypes';
 
 function getData() {
   return fetch('/api/agents')
