+++ client/modules/Dashboard/Dashboard.jsx
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
-import { withRouter } from "react-router";
+import { withRouter } from 'react-router';
 import TemperaturePanel from './TemperaturePanel';
 import SoundPanel from './SoundPanel';
 import CurrentPanel from './CurrentPanel';
