+++ client/modules/Dashboard/TemperaturePanel/index.jsx
@@ -1,17 +1,16 @@
 import { connect } from 'react-redux';
 import TemperaturePanel from './TemperaturePanel';
 import { getTicks } from '../queries';
-import * as selectors from '../../../models/agents/selectors';
 import * as queries from '../../../models/agents/queries';
 
 const mapStateToProps = (state, ownProps) => {
   const { agent } = ownProps;
-  const temperatures = selectors.getTemperatures(agent);
-  const times = selectors.getTimes(agent);
-  const tempArray = queries.getTicks(times, temperatures);
+  const tmpArr = queries.getTemperatures(agent);
+  const timeArr = queries.getTimes(agent);
+  const temperatures = queries.getTicks(timeArr, tmpArr);
 
   return {
-    temperatures: tempArray,
+    temperatures,
   };
 };
