+++ client/modules/Dashboard/index.js
@@ -5,15 +5,15 @@ import * as queries from '../../models/agents/queries';
 const mapStateToProps = (state, ownProps) => {
   const {
     match: {
-      params: { agent }
-      },
+      params: { agent },
+    },
   } = ownProps;
   const agentId = agent;
 
   return {
     error: state.error,
     agent: queries.getAgentById(state, agentId),
-  }
+  };
 };
 
 const mapDispatchToProps = () => ({});
